move lock inside TLCData

+++ src-tauri/src/controller/data/solve.rs
@@ -2,8 +2,8 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub enum IterationMethod {
-    NewtonTangent { h0: f32, max_iter_num: usize },
-    NewtonDown { h0: f32, max_iter_num: usize },
+    NewtonTangent { h0: f64, max_iter_num: usize },
+    NewtonDown { h0: f64, max_iter_num: usize },
 }
 
 impl Default for IterationMethod {
